skill: move pr.lua into skill module
Move lib/build/pr.lua to lib/skill/pr.lua to keep the PR update logic
colocated with the skill dispatcher that uses it.

Changes:
- lib/skill/pr.lua - moved from lib/build/pr.lua
- lib/skill/init.lua - require('skill.pr') instead of require('build.pr')
- lib/cosmic/cook.mk - remove build/pr.lua dependency, included via skill_libs
- Makefile - remove o/lib/build rule, update fallback path to lib/skill/pr.lua
